Get Fantasy Premier League gameweek information.

This function retrieves information about the gameweeks and its details, such as the gameweek name, deadline time, finished status, current status, previous status, next status, highest score, average entry score, transfers made, chip plays, most selected player, most captained player, most vice-captained player, most transferred-in player, and the top player of the gameweek, along with their details.

Usage

fpl_gameweek_info()

Value

A tibble (data frame) with the following columns:

gameweek

The name of the gameweek.

gw_deadline

The deadline datetime of the gameweek in UTC.

gw_finished

A logical value indicating whether the gameweek is finished.

data_checked

A logical value indicating whether the data for the gameweek was checked.

gw_current

A logical value indicating whether the gameweek is the active one.

gw_previous

A logical value indicating whether the gameweek is the previous one.

gw_next

A logical value indicating whether the gameweek is the next one.

gw_highest_score

The highest score achieved by any FPL player in the gameweek.

gw_average_score

The average score of all FPL entries in the gameweek.

transfers_made

The total number of transfers made in the gameweek.

chip_plays

The chip plays made in the gameweek.

player_most_selected

The player who is most selected by FPL managers in the gameweek.

player_most_captained

The player who is most captained by FPL managers in the gameweek.

player_most_vice_captained

The player who is most vice-captained by FPL managers in the gameweek.

player_most_transferred_in

The player who is most transferred-in by FPL managers in the gameweek.

top_player

The top-performing player in the gameweek.

top_player_info

Additional details about the top-performing player of that gameweek.

Examples

fpl_gameweek_info()
#> # A tibble: 38 × 17
#>    gameweek  gw_deadline         gw_finished data_checked gw_current gw_previous
#>    <chr>     <dttm>              <lgl>       <lgl>        <lgl>      <lgl>      
#>  1 Gameweek… 2023-08-11 17:30:00 TRUE        TRUE         FALSE      FALSE      
#>  2 Gameweek… 2023-08-18 17:15:00 TRUE        TRUE         FALSE      FALSE      
#>  3 Gameweek… 2023-08-25 17:30:00 TRUE        TRUE         FALSE      TRUE       
#>  4 Gameweek… 2023-09-01 17:30:00 TRUE        TRUE         TRUE       FALSE      
#>  5 Gameweek… 2023-09-16 10:00:00 FALSE       FALSE        FALSE      FALSE      
#>  6 Gameweek… 2023-09-23 12:30:00 FALSE       FALSE        FALSE      FALSE      
#>  7 Gameweek… 2023-09-30 10:00:00 FALSE       FALSE        FALSE      FALSE      
#>  8 Gameweek… 2023-10-07 10:00:00 FALSE       FALSE        FALSE      FALSE      
#>  9 Gameweek… 2023-10-21 10:00:00 FALSE       FALSE        FALSE      FALSE      
#> 10 Gameweek… 2023-10-27 17:30:00 FALSE       FALSE        FALSE      FALSE      
#> # ℹ 28 more rows
#> # ℹ 11 more variables: gw_next <lgl>, gw_highest_score <int>,
#> #   gw_average_score <int>, transfers_made <int>, chip_plays <list>,
#> #   player_most_selected <chr>, player_most_captained <chr>,
#> #   player_most_vice_captained <chr>, player_most_transferred_in <chr>,
#> #   top_player <chr>, top_player_points <int>