Get Fantasy Premier League classic league information.

This function retrieves information about a classic league in Fantasy Premier League using the provided league ID. It displays the league standings, including each entry's total points, rank, and other relevant details.

Usage

fpl_league_info(league_id)

Arguments

league_id

The ID of the classic league (numeric).

Value

A tibble (data frame) with the following columns:

fpl_entry_id

The FPL entry ID of the participant.

entry_season_id

The ID of the season for the participant.

entry_name

The name of the participant.

entry_team

The team name of the participant.

gw_total

The total points scored by the participant in the current gameweek.

overall_total

The overall total points scored by the participant.

current_rank

The current rank of the participant in the league.

previous_rank

The previous rank of the participant in the league.

Examples

fpl_league_info(12345)
#> → The standings for Blackburn FF Depot league. Last updated at 22:05 on 21 August 2023.
#> 
#> # A tibble: 38 × 8
#>    fpl_entry_id entry_season_id entry_name     entry_team gw_total overall_total
#>           <int>           <int> <chr>          <chr>         <int>         <int>
#>  1       697754         4391697 michael aspin… Eat my du…       56           141
#>  2      1267845         8110253 Steven Waddin… Pending M…       59           136
#>  3      2339971        15688556 Tim Holmes     Pearlmans…       47           135
#>  4       959367         6053755 Eliza Mercer   Pink Elep…       59           134
#>  5       927139        47811661 paul b         here we go       47           129
#>  6      3977764        73821556 William Domink Willdom          48           128
#>  7      3470595        24141752 Archie Aspina… Barons ar…       40           125
#>  8       422095         2583100 Tony Roscoe    Bruno the…       49           125
#>  9       915042         5768881 Lewis Blake    Blakeys Xi       46           124
#> 10       480148         2948225 Michael Shiel… Saka Pota…       43           124
#> # ℹ 28 more rows
#> # ℹ 2 more variables: current_rank <int>, previous_rank <int>