Get Fantasy Premier League player's fixture-level statistics.

This function retrieves detailed fixture-level statistics for a specific player in the Fantasy Premier League (FPL). You can obtain per-game season statistics including total points, starts, minutes played, goals scored, assists, clean sheets, goals conceded, and more.

Usage

fpl_player_fixture_stats(player_id)

Arguments

player_id

The unique ID of the player in the FPL.

Value

A tibble (data frame) with the following columns:

gameweek

The gameweek number.

game_id

The unique ID of the fixture.

opponent

The opponent team's name.

total_points

Total FPL points earned in the gameweek.

starts

The number of starts in the gameweek.

minutes

The total number of minutes played in the gameweek.

goals_scored

The number of goals scored in the gameweek.

assists

The number of assists provided in the gameweek.

clean_sheets

The number of clean sheets in the gameweek.

goals_conceded

The number of goals conceded in the gameweek.

own_goals

The number of own goals scored in the gameweek.

penalties_saved

The number of penalties saved in the gameweek.

penalties_missed

The number of penalties missed in the gameweek.

yellow_cards

The number of yellow cards received in the gameweek.

red_cards

The number of red cards received in the gameweek.

saves

The number of saves made in the gameweek.

bonus

The bonus points earned in the gameweek.

bps

The Bonus Points System (BPS) score in the gameweek.

influence

The influence score in the gameweek.

creativity

The creativity score in the gameweek.

threat

The threat score in the gameweek.

ict_index

The ICT (Influence, Creativity, Threat) index score in the gameweek.

xg_scored

The expected goals scored in the gameweek.

xa

The expected assists in the gameweek.

xg_involvement

The expected goal involvements in the gameweek.

xg_conceded

The expected goals conceded in the gameweek.

price

The player's price for the gameweek.

selected

The number of FPL managers who have selected the player for their team.

transfers_in

The number of transfers in for the player in the gameweek.

transfers_out

The number of transfers out for the player in the gameweek.

transfers_balance

The net transfers balance for the player in the gameweek.

Examples

fpl_player_fixture_stats(player_id = 1)
#> Balogun's per game season statistics
#> # A tibble: 2 × 31
#>   gameweek game_id opponent total_points starts minutes goals_scored assists
#>      <int>   <int> <chr>           <int>  <int>   <int>        <int>   <int>
#> 1        1       2 NFO                 0      0       0            0       0
#> 2        2      12 CRY                 0      0       0            0       0
#> # ℹ 23 more variables: clean_sheets <int>, goals_conceded <int>,
#> #   own_goals <int>, penalties_saved <int>, penalties_missed <int>,
#> #   yellow_cards <int>, red_cards <int>, saves <int>, bonus <int>, bps <int>,
#> #   influence <chr>, creativity <chr>, threat <chr>, ict_index <chr>,
#> #   xg_scored <chr>, xa <chr>, xg_involvement <chr>, xg_conceded <chr>,
#> #   price <dbl>, selected <int>, transfers_in <int>, transfers_out <int>,
#> #   transfers_balance <int>