Print Premier League Table.

This function prints a table with Premier League teams' performance data, including the number of matches played, wins, draws, losses and points.

Usage

pl_table()

Value

A gt table object displaying the Premier League table.

Examples

pl_table()
#> <div id="acklmwzujf" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#acklmwzujf table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #acklmwzujf thead, #acklmwzujf tbody, #acklmwzujf tfoot, #acklmwzujf tr, #acklmwzujf td, #acklmwzujf th {
#>   border-style: none;
#> }
#> 
#> #acklmwzujf p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #acklmwzujf .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #acklmwzujf .gt_title {
#>   color: #FFFFFF;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #acklmwzujf .gt_subtitle {
#>   color: #FFFFFF;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #acklmwzujf .gt_heading {
#>   background-color: #104E8B;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_col_heading {
#>   color: #FFFFFF;
#>   background-color: #1874CD;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #acklmwzujf .gt_column_spanner_outer {
#>   color: #FFFFFF;
#>   background-color: #1874CD;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #acklmwzujf .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #acklmwzujf .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #acklmwzujf .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #acklmwzujf .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #acklmwzujf .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #acklmwzujf .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #acklmwzujf .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #acklmwzujf .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #acklmwzujf .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #4D4D4D;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #acklmwzujf .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #acklmwzujf .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #acklmwzujf .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #acklmwzujf .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #acklmwzujf .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #acklmwzujf .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #acklmwzujf .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #acklmwzujf .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #acklmwzujf .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #acklmwzujf .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #acklmwzujf .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #acklmwzujf .gt_left {
#>   text-align: left;
#> }
#> 
#> #acklmwzujf .gt_center {
#>   text-align: center;
#> }
#> 
#> #acklmwzujf .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #acklmwzujf .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #acklmwzujf .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #acklmwzujf .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #acklmwzujf .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #acklmwzujf .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #acklmwzujf .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #acklmwzujf .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #acklmwzujf .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #acklmwzujf .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #acklmwzujf .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #acklmwzujf .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" style="table-layout: fixed;; width: 0px" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <colgroup>
#>     <col style="width:160px;"/>
#>     <col style="width:80px;"/>
#>     <col style="width:80px;"/>
#>     <col style="width:80px;"/>
#>     <col style="width:80px;"/>
#>     <col style="width:80px;"/>
#>   </colgroup>
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="6" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Premier League Table</strong></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="Team">Team</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Played">Played</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Win">Win</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Draw">Draw</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Loss">Loss</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Points">Points</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(255,215,0,0.7); color: #000000; font-weight: bold;">Arsenal</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(255,215,0,0.7); color: #000000; font-weight: bold;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(255,215,0,0.7); color: #000000; font-weight: bold;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(255,215,0,0.7); color: #000000; font-weight: bold;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(255,215,0,0.7); color: #000000; font-weight: bold;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(255,215,0,0.7); color: #000000; font-weight: bold;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Aston Villa</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Bournemouth</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Brentford</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Brighton</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Burnley</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Chelsea</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Crystal Palace</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Everton</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Fulham</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Liverpool</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Luton</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Man City</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Man Utd</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Newcastle</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Nott'm Forest</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">Sheffield Utd</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(173,216,230,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">Spurs</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">West Ham</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>     <tr><td headers="Team" class="gt_row gt_left" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">Wolves</td>
#> <td headers="Played" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Win" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Draw" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Loss" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td>
#> <td headers="Points" class="gt_row gt_right" style="background-color: rgba(238,162,173,0.5); color: #000000; font-weight: normal;">0</td></tr>
#>   </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="6">As on the 18 August 23 at 21:10 UTC.Created by {fantasypl}.</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>